+++ bittytax/conv/parsers/__init__.py
@@ -11,6 +11,7 @@ from . import circle
 from . import coinbase
 from . import coinbasepro
 from . import coinfloor
+from . import coinomi
 from . import cointracking
 from . import cryptocom
 from . import cryptopia